added gate for new route

+++ app/Http/Controllers/MusiciansController.php
@@ -8,7 +8,6 @@ use App\Models\Instrument;
 use App\Models\Musician;
 use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Foundation\Http\FormRequest;
-use Illuminate\Http\JsonResponse;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Routing\Redirector;
 use Illuminate\Support\Facades\Gate;
@@ -128,6 +127,8 @@ class MusiciansController extends Controller
 
     public function deletePicture(Musician $musician): RedirectResponse
     {
+        Gate::authorize('manage musicians');
+
         Log::info('Deleting picture of musician after API call', [$musician]);
         if ($musician->picture_filepath != null) {
             Storage::delete($musician->picture_filepath);
